// types/FichaMatricula.d.ts
type FichaMatricula = {
  __rowNum__: number;
  cod_alumno: string;
  ape_paterno: string;
  ape_materno: string;
  nom_alumno: string;
  anio_ingreso: string;
  cod_semestre: string;
  cod_plan: string;
  cod_especialidad: number;
  cod_asignatura: string;
  cod_seccion: number;
  num_ciclo_ano_asig: number;
  des_asignatura: string;
  num_creditaje: number;
  mod_user: string;
  mod_fecha: string;
  rep_plan_actual: number;
  sede_alumno: string;
  rep_x_equiv: number;
  coe_alumno: string;
  count: string;
};

export default FichaMatricula;